fitai_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.27
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.0
python-multipart==0.0.9
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
